Update 11/29
This update largely improves on the scheduling functionality seen in
Servlet.java (under the 4RideServlet project folder). The optimalTour
function has been overhauled to accommodate for more flexible and
extendible input, which proved to be a fairly complex change.
Additionally, in the 4Ride Client project folder, CoreLocation has been
enabled and is used to do any location-based calculations within the
app (namely sending current location to servlet and geocoding).