Update 12/6
Adding adjustments to algorithm to accept more weighting factors when
scheduling vehicles, created a cleaner client app view that responds to
user input in a more intuitive way (menus disappear and reappear based
on user state), and improved unit tests. Also clarified the console
output on the server to better demo the algorithm.