Dispatch Module v1.0
This update adds a dispatch module to the server that shows all vehicle
routing data. The Leaflet.JS mapping library is used to display an
interactive map. Outstanding vehicle routes are color-coded. This
information is retrieved directly from the dispatch server. Further,
the location icons can be clicked to show additional information such
as driver name and location address. This update makes use of the
Leaflet Routing Machine — a 3rd party leaflet plug in.