Dispatch Module 2.0, Various Fixes and Improvements
This week largely focused on bug fixes and improvements in preparation
for Wednesday’s presentation. The Dispatch Module has been upgraded to
include vehicle locations (and a custom div icon of a van to better
identify the vehicle location), automatic updating every 5 seconds, and
a new geocoder (MapQuest) to fix an issue where the number of queries
was triggering a MAX_QUERY_LIMIT flag from the previous service. The
student app has also been improved to fix an issue where the map was
not reloading properly.